Fix the build on case-insensitive filesystems and the build timestamp
NwbPage/types.ts and the NwbPage/Types directory collided on macOS and
Windows, where tsc -b resolved ../Types to the file and failed with
TS1149. The overview types now live in Types/index.ts and the file is
gone. The build-info script takes its timestamp from node, since BSD
date printed %3N literally and the site showed an invalid build date.
CONTRIBUTING recommends tsc -b, which the solution-style tsconfig needs,
and the README points at main-v1 for the previous version.

Fixes #496
