+++ devel/generate_build_info.sh
@@ -3,7 +3,9 @@
 SCRIPT_DIR="$(dirname "$0")"
 
 # Generate build-info.json with current timestamp and version
-BUILD_TIME=$(date -u +"%Y-%m-%dT%H:%M:%S.%3NZ")
+# Use node for the timestamp: BSD date (macOS) has no %N, so the GNU format
+# produced a literal "3N" there and the site showed "Built: Invalid Date".
+BUILD_TIME=$(node -p "new Date().toISOString()")
 VERSION=$(node -p "require('$SCRIPT_DIR/../package.json').version")
 BUILD_INFO="{
   \"buildTime\": \"$BUILD_TIME\",
